// src/entities/Profile/model/types/profileSchema.ts
import { Country, Currency } from 'shared/const/common';

// тип профиля
export interface Profile {
    first: string
    lastname: string
    age: number
    currency: Currency
    country: Country
    city: string
    username: string,
    avatar: string
}

// тип который уйдет в стейт
export interface ProfileSchema {
    data?: Profile
    isLoading: boolean
    error?: string
    readonly: boolean
}